fix(importDataSources): set aside config.JSONs for later application
