+++ src/io/import/common.ts
@@ -3,11 +3,13 @@ import { DataSource, FileSource } from '@/src/io/import/dataSource';
 import { Handler } from '@/src/core/pipeline';
 import { ARCHIVE_FILE_TYPES } from '@/src/io/mimeTypes';
 import { Awaitable } from '@vueuse/core';
+import { Config } from '@/src/io/import/configSchema';
 
 export interface ImportResult {
   dataSource: DataSource;
   dataID?: string;
   dataType?: 'image' | 'dicom' | 'model';
+  config?: Config;
 }
 
 export type ArchiveContents = Record<string, File>;
