+++ src/io/import/importDataSources.ts
@@ -1,11 +1,7 @@
-import Pipeline, {
-  PipelineResult,
-  PipelineResultError,
-  PipelineResultSuccess,
-  partitionResults,
-} from '@/src/core/pipeline';
-import { ImportHandler, ImportResult } from '@/src/io/import/common';
+import Pipeline from '@/src/core/pipeline';
+import { ImportHandler } from '@/src/io/import/common';
 import { DataSource, DataSourceWithFile } from '@/src/io/import/dataSource';
+import { nonNullable } from '@/src/utils';
 import handleDicomFile from '@/src/io/import/processors/handleDicomFile';
 import downloadUrl from '@/src/io/import/processors/downloadUrl';
 import extractArchive from '@/src/io/import/processors/extractArchive';
@@ -19,13 +15,7 @@ import updateFileMimeType from '@/src/io/import/processors/updateFileMimeType';
 import handleConfig from '@/src/io/import/processors/handleConfig';
 import { useDICOMStore } from '@/src/store/datasets-dicom';
 import { makeDICOMSelection, makeImageSelection } from '@/src/store/datasets';
-import {
-  EARLIEST_PRIORITY,
-  earliestPriority,
-  prioritizeJSON,
-  prioritizeStateFile,
-  PriorityResult,
-} from './processors/prioritize';
+import { applyConfig } from '@/src/io/import/configSchema';
 
 /**
  * Tries to turn a thrown object into a meaningful error string.
@@ -50,92 +40,57 @@ export async function importDataSources(dataSources: DataSource[]) {
     dicomDataSources: [] as DataSourceWithFile[],
   };
 
-  const prioritizes = [
+  const middleware = [
     // updating the file type should be first in the pipeline
     updateFileMimeType,
-    // before extractArchive as .zip file is part of state file check
-    prioritizeStateFile,
-
-    // source generators
+    // before extractArchive as .zip extension is part of state file check
+    restoreStateFile,
     handleRemoteManifest,
     handleGoogleCloudStorage,
     handleAmazonS3,
     downloadUrl,
     extractArchiveTargetFromCache,
     extractArchive,
-
-    // sinks
-    prioritizeJSON,
-    earliestPriority,
-  ];
-  const prioritizer = new Pipeline<
-    DataSource,
-    PriorityResult | ImportResult,
-    typeof importContext
-  >(prioritizes);
-
-  const priorityOrImportResult = await Promise.all(
-    dataSources.map((r) => prioritizer.execute(r, importContext))
-  );
-
-  const [fetchSuccess, errorOnDataFetches] = partitionResults(
-    priorityOrImportResult
-  );
-
-  const withPriority = fetchSuccess
-    .flatMap((r) => r.data)
-    .map((importOrPriority) => {
-      return {
-        dataSource: importOrPriority.dataSource,
-        priority:
-          'priority' in importOrPriority
-            ? importOrPriority.priority
-            : EARLIEST_PRIORITY,
-      };
-    });
-
-  // Group same priorities, lowest priority first
-  const priorityBuckets = withPriority.reduce((buckets, result) => {
-    const { priority = 0 } = result;
-    if (!buckets[priority]) {
-      // eslint-disable-next-line no-param-reassign
-      buckets[priority] = [];
-    }
-    buckets[priority].push(result.dataSource);
-    return buckets;
-  }, [] as Array<Array<DataSource>>);
-
-  const loaders = [
-    restoreStateFile,
-    handleConfig,
+    handleConfig, // collect to apply later
     // should be before importSingleFile, since DICOM is more specific
     handleDicomFile,
     importSingleFile,
     // catch any unhandled resource
     unhandledResource,
   ];
-  const fileLoader = new Pipeline(loaders);
+  const loader = new Pipeline(middleware);
 
-  // Serially processes buckets by priority.
-  // Configuration in later priority buckets will override earlier ones.
-  const results = [] as Array<PipelineResult<DataSource, ImportResult>>;
-  for (let i = 0; i < priorityBuckets.length; i++) {
-    const sources = priorityBuckets[i];
-    // sparse array, so check for undefined
-    if (sources) {
-      // eslint-disable-next-line no-await-in-loop
-      const bucketResults = await Promise.all(
-        sources.map((dataSource) =>
-          fileLoader.execute(dataSource, importContext)
+  const results = await Promise.all(
+    dataSources.map((r) => loader.execute(r, importContext))
+  );
+
+  // Apply config.JSONs
+  const configResult = await (async () => {
+    try {
+      results
+        .flatMap((pipelineResult) =>
+          pipelineResult.ok ? pipelineResult.data : []
         )
-      );
-      results.push(...bucketResults);
+        .map((result) => result.config)
+        .filter(nonNullable)
+        .forEach(applyConfig);
+      return {
+        ok: true as const,
+        data: [],
+      };
+    } catch (err) {
+      return {
+        ok: false as const,
+        errors: [
+          {
+            message: toMeaningfulErrorString(err),
+            cause: err,
+            inputDataStackTrace: [],
+          },
+        ],
+      };
     }
-  }
-
-  if (!importContext.dicomDataSources.length) {
-    return [...errorOnDataFetches, ...results];
-  }
+  })();
 
   // handle DICOM loading
   const dicomDataSource: DataSource = {
@@ -145,20 +100,29 @@ export async function importDataSources(dataSources: DataSource[]) {
   };
   const dicomResult = await (async () => {
     try {
+      if (!importContext.dicomDataSources.length) {
+        return {
+          ok: true as const,
+          data: [],
+        };
+      }
       const volumeKeys = await useDICOMStore().importFiles(
         importContext.dicomDataSources
       );
       return {
-        ok: true,
-        data: volumeKeys.map((key) => ({
-          dataID: key,
-          dataType: 'dicom' as const,
-          dataSource: dicomDataSource,
-        })),
-      } as PipelineResultSuccess<ImportResult>;
+        ok: true as const,
+        data: volumeKeys.map(
+          (key) =>
+            ({
+              dataID: key,
+              dataType: 'dicom' as const,
+              dataSource: dicomDataSource,
+            } as const)
+        ),
+      };
     } catch (err) {
       return {
-        ok: false,
+        ok: false as const,
         errors: [
           {
             message: toMeaningfulErrorString(err),
@@ -166,16 +130,16 @@ export async function importDataSources(dataSources: DataSource[]) {
             inputDataStackTrace: [dicomDataSource],
           },
         ],
-      } as PipelineResultError<typeof dicomDataSource>;
+      };
     }
   })();
 
   return [
-    ...errorOnDataFetches,
     // remove all results that have no result data
-    ...results.filter((result) => !result.ok || result.data.length),
+    ...results,
     dicomResult,
-  ];
+    configResult,
+  ].filter((result) => !result.ok || (result.ok && result.data.length));
 }
 
 export type ImportDataSourcesResult = Awaited<
